requirements.txt
pymongo==4.6.1
mysql-connector-python==8.3.0
paho-mqtt==2.1.0